requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0
openai>=1.40.0
